+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-# 0.7.0
+# 1.0
 
-- initial release as standalone package
+- stable release as standalone package
+- add download_dist function directly into CLI tool, instead of API
+- convert exit control flow to a main function return-based method